requirements.txt
# Flask Web Framework
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# TensorFlow for ML Model
tensorflow-cpu>=2.15.0

# Image Processing
pillow>=10.2.0
numpy>=1.26.4
opencv-python-headless>=4.9.0.80
